Exclude unix: from vnc call to unix_listen() (Ryan Harper)
When using an existing unix socket like:

 -vnc unix:/tmp/file1Y2nY2

qemu fails to bind a unix socket because the vnc call to unix_listen includes
the unix: prefix and stores that in the unix.sun_path.  The fix is to not pass
in unix: for the filename (same way qemu-char.c does it).

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>



git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@5856 c046a42c-6fe2-441c-8c8c-71466251a162

committer: aliguori <aliguori@c046a42c-6fe2-441c-8c8c-71466251a162>
git-commit-id: 4a55bfdfcf051f45619d422285be4b8c30ff20ae