fix usb-hid emulation SET_IDLE
there is a bug in the current usb-hid emulation code that affects the
way the idle flag is handled: the spec clearly states that when SET_IDLE
is called with a value == 0 means "do not send any event unless there is
an actual change". However we are currently assuming the opposite
meaning of the flag.
This patch fixes it and it also fixes a possible initialization issue
that is uncovered by using the default idle flag set to 0.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>
[supplied patch also trivially changed by Ian Jackson]

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: 144cb9b13bb57d40b69a5100ab4dd91d9f0036e0