populate guest videoram
This patch applies to qemu-xen-unstable and implements the new way of
populating the videoram in the guest and mapping it in qemu.  Initially
the videoram is populated at the address 0xff000000 in the guest memory
address space.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>

(This is Stefano's [PATCH 2 of 4] ioemu;
 1 is in xen-unstable and 3 and 4 will go in shortly -iwj.)

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: 8ef3508bba665f14714274ce5504222924335555