hw/passthrough.[ch], pt-msi.[ch]: remove trailing whitespace
Trailing whitespace does not survive some people's mailers, with the
result that their patches do not apply.  Therefore in this commit I'm
removing it from the pass-through code, which is currently entirely in
the Xen tree.

This change is the result of this rune:
 perl -i~ -pe 's/[\t ]*$//' hw/pass-through.[ch] hw/pt-msi.[ch]

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: d42def00dad5cb5bdf2d5a6b10a079cea6060216